Fix code smell ClassFileName
Class name "Categories" doesn't match filename, expected "universe_create_locations"
(Spryker.Classes.ClassFileName.NoMatch)

To fix this, we move the `Categories` class out of the engine file and
into a new `Smr\Admin` namespace. It doesn't really make sense outside
the context of the engine file, but there's really no other logical
place to put it.
