+++ src/engine/Default/admin/unigen/universe_create_locations.php
@@ -1,5 +1,7 @@
 <?php declare(strict_types=1);
 
+use Smr\Admin\UniGenLocationCategories;
+
 $template = Smr\Template::getInstance();
 $session = Smr\Session::getInstance();
 $var = $session->getCurrentVar();
@@ -30,31 +32,6 @@ foreach ($galaxy->getSectors() as $galSector) {
 }
 $template->assign('TotalLocs', $totalLocs);
 
-// Though we expect a location to be only in one category, it is possible to
-// edit a location in the Admin Tools so that it is in two or more categories.
-// For simplicity here, it will only show up in the first category it matches,
-// but it will identify all other categories that it is in.
-// If multi-category locations becomes common, this code should be modified.
-class Categories {
-
-	/** @var array<string, array<int>> */
-	public array $locTypes = [];
-	/** @var array<int> */
-	private array $locAdded = []; // list of locs added to a category
-	public function addLoc(int $locID, string $category): string {
-		if ($this->added($locID)) {
-			return "<b>Also in $category</b><br />";
-		}
-		$this->locTypes[$category][] = $locID;
-		$this->locAdded[] = $locID;
-		return '';
-	}
-	public function added(int $locID): bool {
-		return in_array($locID, $this->locAdded);
-	}
-
-}
-
 // Remove any linked locations, as they will be added automatically
 // with any corresponding HQs.
 foreach ($locations as $location) {
@@ -65,7 +42,7 @@ foreach ($locations as $location) {
 
 // Set any extra information to be displayed with each location
 $locText = [];
-$categories = new Categories();
+$categories = new UniGenLocationCategories();
 foreach ($locations as $location) {
 	$extra = '<span class="small"><br />';
 	if ($location->isWeaponSold()) {
